fgsdfgsdfgs

+++ resources/views/admin/users/show.blade.php
@@ -11,7 +11,9 @@
                     <div class="card-body">
 
                         <a href="{{ url('/admin/users') }}" title="Back"><button class="btn btn-warning btn-sm"><i class="fa fa-arrow-left" aria-hidden="true"></i> Back</button></a>
+                @if (Auth::user()->hasRole('superadmin'))         
                         <a href="{{ url('/admin/users/' . $user->id . '/edit') }}" title="Edit User"><button class="btn btn-primary btn-sm"><i class="fa fa-pencil-square-o" aria-hidden="true"></i> Edit</button></a>
+                     
                         {!! Form::open([
                             'method' => 'DELETE',
                             'url' => ['/admin/users', $user->id],
@@ -24,6 +26,7 @@
                                     'onclick'=>'return confirm("Confirm delete?")'
                             ))!!}
                         {!! Form::close() !!}
+                 @endif
                         <br/>
                         <br/>
